kp2

+++ src/main/java/de/uni_mannheim/informatik/dws/wdi/ExerciseIdentityResolution/Comparators/author/BookAuthorComparatorJaccard.java
@@ -49,7 +49,7 @@ public class BookAuthorComparatorJaccard implements Comparator<Book, Attribute> 
 		}
 		
 		// postprocessing
-		double postSimilarity = 1;
+		int postSimilarity = 1;
 		if (maxSimilarity <= 0.3) {
 			postSimilarity = 0;
 		}
