+++ src/main/java/de/uni_mannheim/informatik/dws/wdi/ExerciseIdentityResolution/Comparators/author/BookAuthorComparatorJaroWinkler.java
@@ -48,7 +48,7 @@ public class BookAuthorComparatorJaroWinkler implements Comparator<Book, Attribu
         }
 
         // postprocessing
-        double postSimilarity = 1;
+        int postSimilarity = 1;
         if (maxSimilarity <= 0.3) {
             postSimilarity = 0;
         }
